Repos update

+++ src/repository/TransactionRepo.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 public interface TransactionRepo {
     void saveTransaction(Transaction transaction);
-    void deleteTransaction(Long id, Transaction transaction);
-    Transaction findTransactionById(Long id);
-    Map<Long, Transaction> findAllTransactions();
+    void deleteTransaction(int id, Transaction transaction);
+    Transaction findTransactionById(int id);
+    Map<Integer, Transaction> findAllTransactions();
 }
